requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
# pandas==2.1.3
# openai==1.3.5
openai==1.12.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0